# Conflicts:
#	src/main/java/com/example/retro_care/customer/controller/CustomerController.java
#	src/main/java/com/example/retro_care/customer/dto/CustomerDto.java
#	src/main/java/com/example/retro_care/customer/repository/ICustomerRepository.java
#	src/main/java/com/example/retro_care/customer/service/CustomerService.java
#	src/main/java/com/example/retro_care/customer/service/ICustomerService.java
#	src/main/java/com/example/retro_care/employee/controller/EmployeeController.java
#	src/main/java/com/example/retro_care/employee/repository/IEmployeeRepository.java
#	src/main/java/com/example/retro_care/employee/service/EmployeeService.java
#	src/main/java/com/example/retro_care/indication/controller/IndicationController.java
#	src/main/java/com/example/retro_care/indication/service/IIndicationService.java
#	src/main/java/com/example/retro_care/medicine/controller/MedicineController.java
#	src/main/java/com/example/retro_care/medicine/repository/IMedicineRepository.java
#	src/main/java/com/example/retro_care/medicine/service/IMedicineService.java
#	src/main/java/com/example/retro_care/medicine/service/Impl/MedicineService.java
#	src/main/java/com/example/retro_care/prescription/controller/PrescriptionController.java
#	src/main/java/com/example/retro_care/prescription/model/Prescription.java
#	src/main/java/com/example/retro_care/prescription/service/PrescriptionService.java
#	src/main/java/com/example/retro_care/user/common/ValidateAppUser.java
#	src/main/java/com/example/retro_care/user/controller/AppUserController.java
#	src/main/java/com/example/retro_care/user/repository/IAppUserRepository.java
#	src/main/java/com/example/retro_care/user/service/IAppUserService.java

+++ src/main/java/com/example/retro_care/employee/controller/EmployeeController.java
@@ -3,27 +3,36 @@ package com.example.retro_care.employee.controller;
 
 import com.example.retro_care.employee.dto.EmployeeDto;
 import com.example.retro_care.employee.model.Employee;
+import com.example.retro_care.user.common.ValidateAppUser;
+import com.example.retro_care.user.model.AppUser;
+import com.example.retro_care.user.service.IAppUserService;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.validation.BindingResult;
 import com.example.retro_care.employee.service.IEmployeeService;
-import org.apache.coyote.Request;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
 
+import org.springframework.validation.FieldError;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.HashMap;
+import java.util.Map;
 @RestController
 @CrossOrigin("*")
-@RequestMapping("/employees")
+@RequestMapping("/api/employees")
 public class EmployeeController {
     @Autowired
     private IEmployeeService employeeService;
-
+    @Autowired
+    private IAppUserService appUserService;
+    @Autowired
+    private PasswordEncoder passwordEncoder;
     /**
      * Author: TanNV
      * Date: 15/09/2023
@@ -35,6 +44,7 @@ public class EmployeeController {
     public ResponseEntity<EmployeeDto> getEmployeeToCreate() {
         EmployeeDto employeeDto = new EmployeeDto();
         employeeDto.setCodeEmployee(employeeService.getNextCode());
+        employeeDto.setImage("https://i.bloganchoi.com/bloganchoi.com/wp-content/uploads/2022/02/avatar-trang-y-nghia.jpeg?fit=512%2C20000&quality=95&ssl=1");
         return new ResponseEntity<>(employeeDto, HttpStatus.OK);
     }
 
@@ -44,34 +54,73 @@ public class EmployeeController {
      * Receive data and validate, if there is an error, return BAD_REQUEST,
      * then save the employee to the DB. If saved successfully, return OK, otherwise NO_CONTENT
      *
-     * @param employeeDto
-     * @param bindingResult
+     * @param employeeDto validate
+     * @param bindingResult errors
      * @return Response entity
      */
     @PostMapping("/create")
-    public ResponseEntity<String> createEmployee(@RequestBody EmployeeDto employeeDto, BindingResult bindingResult) {
+    public ResponseEntity<?> createEmployee(@RequestBody EmployeeDto employeeDto, BindingResult bindingResult) {
         new EmployeeDto().validate(employeeDto, bindingResult);
+        Map<String, String> errorMap= new HashMap<>();
         if (bindingResult.hasErrors()) {
-            return new ResponseEntity<>(bindingResult.getAllErrors().toString(),HttpStatus.BAD_REQUEST);
+            for (FieldError fieldError: bindingResult.getFieldErrors()
+                 ) {
+                errorMap.put(fieldError.getField(),fieldError.getDefaultMessage());
+            }
+            return new ResponseEntity<>(errorMap,HttpStatus.NOT_ACCEPTABLE);
         }
+        Employee employeeCheck = employeeService.getByPhoneNumber(employeeDto.getPhoneNumber(),-1L);
+        if(employeeCheck!=null){
+            errorMap.put("phoneNumber","SĐT đã được đăng ký");
+        }
+
+        String errMsg = ValidateAppUser.checkValidateOnlyAppUserName(employeeDto.getAppUser());
+        if (!errMsg.equals("")) {
+            return ResponseEntity
+                    .status(HttpStatus.NOT_ACCEPTABLE)
+                    .body(errMsg);
+        }
+
+        Boolean userNameExisted = appUserService.existsByUsername(employeeDto.getAppUser());
+        if (Boolean.TRUE.equals(userNameExisted)) {
+            errorMap.put("appUser","Tài khoản này đã tồn tại");
+
+        }
+        if(errorMap.size()>0){
+            return new ResponseEntity<>(errorMap,HttpStatus.NOT_ACCEPTABLE);
+        }
+
+        AppUser appUser = new AppUser();
+        appUser.setUserName(employeeDto.getAppUser());
+        appUser.setPassword(passwordEncoder.encode("123"));
+        Boolean checkAddNewAppUser = appUserService.createNewAppUser(appUser,"ROLE_EMPLOYEE");
+        if (!Boolean.TRUE.equals(checkAddNewAppUser)) {
+            return ResponseEntity
+                    .status(HttpStatus.INTERNAL_SERVER_ERROR)
+                    .body("Đăng ký thất bại, vui lòng chờ trong giây lát");
+        }
+        Long userId = appUserService.findAppUserIdByUserName(employeeDto.getAppUser());
         Employee employee = new Employee();
         BeanUtils.copyProperties(employeeDto, employee);
-        employeeService.addEmployee(employee);
-        return new ResponseEntity<>("Create successfully", HttpStatus.OK);
+        employeeService.addEmployee(employee,userId);
+        return new ResponseEntity<>("Thêm mới thành công", HttpStatus.OK);
     }
 
     /**
      * Author: TanNV
      * Date: 15/09/2023
      * Use to get employee by id and return Http status OK if it can find it else  return http status NO_CONTENT
-     * @param id
+     * @param id employee id
      * @return Reponse entity
      */
     @GetMapping("/{id}")
     public ResponseEntity<Employee> getEmployee(@PathVariable Long id){
+        if(id==null){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         Employee employee = employeeService.getById(id);
         if(employee==null){
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
         return new ResponseEntity<>(employee,HttpStatus.OK);
     }
@@ -86,39 +135,29 @@ public class EmployeeController {
      * @param bindingResult return error
      * @return Responese Entity with message
      */
-    @PutMapping("/update/{id}")
+    @PatchMapping("/update/{id}")
     public ResponseEntity<String> updateEmployee(@PathVariable Long id,
                                                    @RequestBody EmployeeDto employeeDto,
                                                    BindingResult bindingResult){
+        if (id == null){
+            return new ResponseEntity<>("Không có id",HttpStatus.BAD_REQUEST);
+        }
         new EmployeeDto().validate(employeeDto, bindingResult);
         if (bindingResult.hasErrors()) {
-            return new ResponseEntity<>(bindingResult.getAllErrors().toString(),HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>(bindingResult.getAllErrors().toString(),HttpStatus.NOT_ACCEPTABLE);
         }
         Employee employee = employeeService.getById(id);
+        if(employee==null){
+            return new ResponseEntity<>("Không tìm thấy",HttpStatus.NOT_FOUND);
+        }
+        Employee employeeCheck = employeeService.getByPhoneNumber(employeeDto.getPhoneNumber(),id);
+        if(employeeCheck!=null){
+            return new ResponseEntity<>("SĐT đã được đăng ký",HttpStatus.NOT_ACCEPTABLE);
+        }
+
         BeanUtils.copyProperties(employeeDto, employee);
         employeeService.updateEmployee(employee);
-        return new ResponseEntity<>("Update successfully",HttpStatus.OK);
-    }
-    /**
-     * Create: SonTT
-     * Date create: 15/09/2023
-     * Function: Call the database to retrieve the data with page, limit and sort
-     *
-     * @param page
-     * @param limit
-     * @return ResponseEntity<?>
-     */
-    @GetMapping("/get-list/{page}/{limit}/{sort}")
-    public ResponseEntity<Page<Employee>> getListEmployee(@PathVariable(value = "page", required = false) Integer page,
-                                                          @PathVariable(value = "limit", required = false) Integer limit,
-                                                          @PathVariable(value = "sort", required = false) String sort) {
-        Pageable pageable = PageRequest.of(page, limit);
-        Page<Employee> employees = employeeService.getListEmployee(pageable);
-        if (employees.isEmpty()) {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-        } else {
-            return new ResponseEntity<>(employees, HttpStatus.OK);
-        }
+        return new ResponseEntity<>("Update thành công",HttpStatus.OK);
     }
 
 
@@ -126,44 +165,51 @@ public class EmployeeController {
      * Create: SonTT
      * Date create: 15/09/2023
      * Function: Call the database to retrieve paginated data with fields idRole and employee name
-     *
-     * @param page
-     * @param limit
-     * @param sort
-     * @param idRole
-     * @param nameEmployee
+     * @param page number of page
+     * @param limit limit element in page
+     * @param sort sort
+     * @param nameEmployee name employee
      * @return ResponseEntity<?>
      */
-    @GetMapping("/search-list/{page}/{limit}")
+    @GetMapping("/list/{page}/{limit}/{sort}")
     public ResponseEntity<Page<Employee>> searchEmployee(@PathVariable(value = "page", required = false) Integer page,
                                                          @PathVariable(value = "limit", required = false) Integer limit,
                                                          @PathVariable(value = "sort", required = false) String sort,
-                                                         @RequestParam(value = "role", required = false) Long idRole,
                                                          @RequestParam(value = "name", required = false) String nameEmployee) {
-        Pageable pageable = PageRequest.of(page, limit, Sort.by(sort));
-        Page<Employee> employees = employeeService.searchEmployee(pageable, idRole, nameEmployee);
+        Pageable pageable = PageRequest.of(page, limit, Sort.by(Sort.Direction.ASC, sort));
+        Page<Employee> employees;
+        if (nameEmployee == null) {
+             employees = employeeService.getListEmployee(pageable);
+
+        } else {
+             employees = employeeService.searchEmployee(pageable, nameEmployee);
+        }
         if (employees.isEmpty()) {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         } else {
             return new ResponseEntity<>(employees, HttpStatus.OK);
         }
-
     }
 
     /**
      * Create:SonTT
      * Date create: 15/09/2023
      * Function: with the correct input parameter id true then return HttpStatus.OK otherwise return false
-     * @param id
+     * @param id id employee
      * @return ResponseEntity<>
      */
-    @DeleteMapping("/delete")
+    @DeleteMapping("/delete-employee")
     public ResponseEntity<HttpStatus> deleteEmployee(@RequestParam(value = "id", required = false) Long id) {
-        if (employeeService.deleteEmployee(id)) {
-            return new ResponseEntity<>(HttpStatus.OK);
-        } else {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        if (employeeService.findEmployee(id)==null){
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }else {
+            if (employeeService.deleteEmployee(id)) {
+                return new ResponseEntity<>(HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
         }
+
 
     }
